Refactor ui/view to seperate node graph generation logic from neural network logic, and remove any modules and code snippets not relevant to VSE - Remove redundant modules - Refactor structure of ui/view by seperating QT painting logic and verification logic. Add a directory ui/view/verification for storing derived classes with verification logic - Adjust contributions in module docstrings - Change class naming scheme for clear model-view seperation
