--- ui/view/__init__.py
@@ -1,15 +1,0 @@
-from .dummy_types import (
-    Block, Edge, Socket, Scene, PropertyBlock, LayerBlock, Project, 
-    CoCoNetWindow, GraphicsScene, FileFormat, MessageType,
-    read_properties, get_classname, RES_DIR
-)
-
-from .graphics_scene import GraphicsScene as ActualGraphicsScene
-from .graphics_view import GraphicsView
-
-__all__ = [
-    'Block', 'Edge', 'Socket', 'Scene', 'PropertyBlock', 'LayerBlock', 'Project',
-    'CoCoNetWindow', 'GraphicsScene', 'FileFormat', 'MessageType',
-    'read_properties', 'get_classname', 'RES_DIR',
-    'ActualGraphicsScene', 'GraphicsView'
-]
