--- ui/view/component/__init__.py
@@ -1,11 +1,0 @@
-from .graphics_block import GraphicsBlock
-from .graphics_edge import GraphicsEdge, GraphicsDirectEdge, GraphicsBezierEdge
-from .graphics_socket import GraphicsSocket
-from .inspector import InspectorDockToolbar
-
-__all__ = [
-    'GraphicsBlock',
-    'GraphicsEdge', 'GraphicsDirectEdge', 'GraphicsBezierEdge',
-    'GraphicsSocket',
-    'InspectorDockToolbar'
-]
