+++ ui/view/graphics_view.py
@@ -3,7 +3,7 @@ Module graphics_view.py
 
 This module contains the GraphicsView class for rendering graphics objects in the viewport
 
-Author: Andrea Gimelli, Giacomo Rosato, Stefano Demarchi
+Original author: Andrea Gimelli, Giacomo Rosato, Stefano Demarchi
 
 """
 
@@ -13,26 +13,7 @@ from PyQt6 import QtGui
 from PyQt6.QtWidgets import QGraphicsView
 
 import ui.view.styling.dimension as dim
-
-# Import from dummy_types to resolve type annotations
-from .dummy_types import GraphicsScene
-
-# Dummy classes to replace imports
-class LayerBlock:
-    pass
-
-class PropertyBlock:
-    pass
-
-class MessageDialog:
-    def __init__(self, message, message_type):
-        pass
-    
-    def exec(self):
-        pass
-
-class MessageType:
-    ERROR = "error"
+from .graphics_scene import GraphicsScene
 
 
 class GraphicsView(QGraphicsView):
@@ -83,43 +64,10 @@ class GraphicsView(QGraphicsView):
         if not clipped:
             self.scale(factor, factor)
 
-    def check_delete(self):
-        """
-        This method checks the selected items in the scene and tries to delete them
-
-        """
-
-        delete = True
-        sel_ids = []
-
-        for item in self.gr_scene_ref.selectedItems():
-            if hasattr(item, 'block_ref'):
-                if isinstance(item.block_ref, LayerBlock):
-                    sel_ids.append(item.block_ref.id)
-
-        for i in range(len(sel_ids)):
-            if self.gr_scene_ref.scene_ref.sequential_list[-2 - i] not in sel_ids:
-                delete = False
-
-        if delete:
-            self.delete_items(sel_ids)
-        else:
-            dialog = MessageDialog('It is only allowed to delete blocks at the end of the network',
-                                   MessageType.ERROR)
-            dialog.exec()
-
     def delete_items(self, sel_ids: list):
         for block_id in sel_ids:
             block = self.gr_scene_ref.scene_ref.blocks[block_id]
             self.gr_scene_ref.scene_ref.remove_block(block, logic=True)
-
-        for item in self.gr_scene_ref.selectedItems():
-            if hasattr(item, 'block_ref'):
-                if isinstance(item.block_ref, PropertyBlock):
-                    if item.block_ref.ref_block.title == 'Input':
-                        self.gr_scene_ref.scene_ref.remove_in_prop()
-                    else:
-                        self.gr_scene_ref.scene_ref.remove_out_prop()
 
     def mousePressEvent(self, event: 'QtGui.QMouseEvent') -> None:
         """
