+++ ui/view/styling/__init__.py
@@ -1,17 +1,13 @@
 """
 Styling module for the node editor interface.
 
-This module provides styling constants, dimensions, color palette, and custom widgets.
+This module provides styling constants, dimensions, and color palette.
 
-Author: Migration Assistant
 """
 
 from . import dimension
 from . import palette
-from . import display
-from .custom import CustomButton, CustomLabel, CustomTextBox, CustomComboBox, CustomTextArea
 
 __all__ = [
-    'dimension', 'palette', 'display',
-    'CustomButton', 'CustomLabel', 'CustomTextBox', 'CustomComboBox', 'CustomTextArea'
+    'dimension', 'palette'
 ]
